requirements.txt
streamlit
pillow
imagehash
numpy
pandas
opencv-python-headless
rembg
streamlit-webrtc
av